requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
pdfplumber
pandas
numpy
xgboost
scikit-learn
tensorflow-cpu
shap
requests
matplotlib
seaborn
scipy
